Revert "use u64 hash"
This reverts commit 6f2753eec9ac801199118ca8d597493c8965bb0d.

+++ crates/rspack_ids/src/deterministic_chunk_ids_plugin.rs
@@ -105,9 +105,9 @@ async fn chunk_ids(&self, compilation: &mut rspack_core::Compilation) -> rspack_
       chunk_key_to_id.insert(chunk.ukey(), id);
       true
     },
-    &[u64::pow(10, max_length)],
+    &[usize::pow(10, max_length)],
     expand_factor,
-    used_ids_len as u64,
+    used_ids_len,
     salt,
   );
